+++ app/src/main/res/layout/fragment_settings_bottom_sheet.xml
@@ -14,6 +14,7 @@
 
     <TextView
         android:id="@+id/bsheet_textView"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="16dp"
         android:text="user_nickname"
