+++ app/src/main/res/values/string-array.xml
@@ -1,33 +1,33 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string-array name="TIPS">
-        <item> 재활용(유리병, 종이팩, 금속캔, 페트병) 시 내용물과 이물질은 제거하고 배출해주세요.<br></br>
+        <item> 재활용(유리병, 종이팩, 금속캔, 페트병) 시 내용물과 이물질은 제거하고 배출해주세요.\n
             이물질이 남아있거나 국물이 스며들어 남아있다면 그 제품은 재활용될 수 없습니다.</item>
-        <item>종이류와 종이팩은 구분해서 버려주세요.<br></br>
+        <item>종이류와 종이팩은 구분해서 버려주세요.\n
             종이류와 종이팩은 재활용 공정의 차이로 인해 종이류는 새 종이로, 종이팩은 화장지, 미용티슈와 같은 서로 다른 제품으로 재탄생되기 때문입니다.</item>
-        <item>새 옷을 사는 대신 친구들과 교환해주세요.<br></br>
-            패스트패션 브랜드 대신 구제샵에서 옷을 사는 것도 좋습니다.<br></br>
+        <item>새 옷을 사는 대신 친구들과 교환해주세요.\n
+            패스트패션 브랜드 대신 구제샵에서 옷을 사는 것도 좋습니다.\n
             전 세계 온실가스 배출량의 10%, 폐수 발생의 20%를 의류 업계가 차지하고 있기 때문에 이를 방지하기 위해 교환하는 습관이 필요합니다.</item>
-        <item>용기 밖의 부착상표 등 타 재질 부분은 제거해주세요.<br></br>
-            페트병의 부착상표는 제거하여 종량제 봉투에 배출하고/ 유리병의 뚜껑은 철, 알루미늄 등 뚜껑의 재질에 맞춰 분리배출해주세요.<br></br>
+        <item>용기 밖의 부착상표 등 타 재질 부분은 제거해주세요.\n
+            페트병의 부착상표는 제거하여 종량제 봉투에 배출하고/ 유리병의 뚜껑은 철, 알루미늄 등 뚜껑의 재질에 맞춰 분리배출해주세요.\n
             병과 병의 외부에 부착된 라벨은 그 재질이 다르기 때문에 병과 라벨을 함께 파쇄하게 되면 제품의 재생 품질이 급격히 떨어져 재활용도가 낮아지는 문제점이 존재합니다.</item>
-        <item>길이가 긴 채소류와 덩어리가 큰 수박 등은 잘게 썰어 부피를 줄이고 물기를 꽉짜서 최대한 수분을 제거한 후 배출해주세요.<br></br>
+        <item>길이가 긴 채소류와 덩어리가 큰 수박 등은 잘게 썰어 부피를 줄이고 물기를 꽉짜서 최대한 수분을 제거한 후 배출해주세요.\n
             음식물류 폐기물은 사료화 및 퇴비화 공정 등을 통해 재활용되며, 바이오가스화를 통해 자동차 연료 등의 에너지로도 재활용됩니다.</item>
-        <item>핵과류의 씨, 껍데기(갑각류, 어패류), 뼈(닭 등의 뼈다귀, 생선뼈), 티백, 한약재, 복어 내장은 쓰레기종량제 봉투로 배출해주세요.<br></br>
+        <item>핵과류의 씨, 껍데기(갑각류, 어패류), 뼈(닭 등의 뼈다귀, 생선뼈), 티백, 한약재, 복어 내장은 쓰레기종량제 봉투로 배출해주세요.\n
             이와 같은 쓰레기는 수분이 적고 딱딱해서 재활용하기 어렵습니다.</item>
-        <item>폐의약품은 신경써서 분리배출해야합니다.<br></br>
-            가정 내 폐의약품이 매립되거나 도시하수로 배출될 경우 공기, 토양, 수질 등의 환경오염을 유발하고 생태계 교란의 원인이 됩니다.<br></br>
+        <item>폐의약품은 신경써서 분리배출해야합니다.\n
+            가정 내 폐의약품이 매립되거나 도시하수로 배출될 경우 공기, 토양, 수질 등의 환경오염을 유발하고 생태계 교란의 원인이 됩니다.\n
             약국, 보건소, 보건지소 또는 보건진료소로 배출해주세요.</item>
     </string-array>
     <string-array name="WARNINGS">
         <item>먹이사슬 최상위 포식자로써 북극 생태계 균형을 유지하는 데 중요한 역할을 하는 북극곰은 2050년에 완전히 사라질 지 모릅니다.</item>
-        <item>세계자연기금(WWF)에 따르면 한 사람이 일주일간 삼키는 미세플라스틱 알갱이는 2000개에 달합니다.<br></br>
+        <item>세계자연기금(WWF)에 따르면 한 사람이 일주일간 삼키는 미세플라스틱 알갱이는 2000개에 달합니다.\n
             무게로 따지면 신용카드 한 장에 해당하는 5g 정도, 한 달 기준으로는 칫솔 무게와 비슷한 21g, 연간으로는 250g에 이릅니다.</item>
-        <item>플라스틱 아일랜드, 쓰레기 섬을 아시나요?<br></br>
+        <item>플라스틱 아일랜드, 쓰레기 섬을 아시나요?\n
             90% 이상이 플라스틱 제품으로 이루어진 이 쓰레기 섬은 2018년 연구에 따르면 크기가 급속하게 불어나 한반도 면적의 무려 7배인 155만㎢에 달한다는 결과가 나왔습니다.</item>
-        <item>연간 인천 앞바다에 버려지는 쓰레기는 연간 19만 ㎥에 달합니다.<br></br>
+        <item>연간 인천 앞바다에 버려지는 쓰레기는 연간 19만 ㎥에 달합니다.\n
             이는 10톤 트럭 1만 여 대에 달하는 쓰레기가 해마다 바다로 버려지고 있음을 뜻합니다.</item>
-        <item>BBC News 코리아에 따르면, ‘지구의 허파’ 아마존의 삼림 벌채가 15년 만에 최대치를 기록한 것으로 나타났습니다.<br></br>
+        <item>BBC News 코리아에 따르면, ‘지구의 허파’ 아마존의 삼림 벌채가 15년 만에 최대치를 기록한 것으로 나타났습니다.\n
             브라질 국립 우주연구소의 보고서에 따르면, 아마존 열대우림의 삼림 벌채는 1년 만에 22% 증가했다고 합니다.</item>
     </string-array>
     <string-array name="CHALLENGES">
