+++ app/src/main/res/values/strings.xml
@@ -30,7 +30,7 @@
     <string name="settings_contact">Contact</string>
     <string name="settings_about_us">About Us</string>
     <string name="settings_logout_dialog">로그아웃 하시겠습니까?</string>
-    <string name="settings_del_dialog">회원탈퇴 진행 시 지금까지의 모든 기록이 삭제됩니다.<br></br>계속 진행할까요?</string>
+    <string name="settings_del_dialog">회원탈퇴 진행 시 지금까지의 모든 기록이 삭제됩니다.\n계속 진행할까요?</string>
     <string name="settings_del_progess">회원탈퇴 진행을 위해 비밀번호를 입력해주세요.</string>
     <string name="password_fail">비밀번호가 올바르지 않습니다.</string>
     <string name="del_comp">회원탈퇴가 완료되었습니다.<br></br>이용해주셔서 감사합니다.</string>
